Modifications to correct handling of G00 commands
G00 defines moves which tell your machine to go as fast as it can to the next location.  The Laser should be turned off during these moves and the feed rate should be restored following the move.  The original implementation did not handle this correctly.

Corresponding changes were made to the Turnkeylaser inkScape extension.  They are included in the fork since they are also required to work with the changes made to the firmware.

+++ Marlin/ultralcd_st7920_u8glib_rrd.h
@@ -21,9 +21,9 @@
 
 #include <U8glib.h>
 
-static void ST7920_SWSPI_SND_8BIT(uint8_t val)
+static void ST7920_SWSPI_SND_8BIT(int val)
 {
-  uint8_t i;
+  int i;
   for( i=0; i<8; i++ )
   {
     WRITE(ST7920_CLK_PIN,0);
@@ -40,9 +40,9 @@ static void ST7920_SWSPI_SND_8BIT(uint8_t val)
 #define ST7920_WRITE_BYTE(a)     {ST7920_SWSPI_SND_8BIT((a)&0xf0);ST7920_SWSPI_SND_8BIT((a)<<4);u8g_10MicroDelay();}
 #define ST7920_WRITE_BYTES(p,l)  {uint8_t i;for(i=0;i<l;i++){ST7920_SWSPI_SND_8BIT(*p&0xf0);ST7920_SWSPI_SND_8BIT(*p<<4);p++;}u8g_10MicroDelay();}
 
-uint8_t u8g_dev_rrd_st7920_128x64_fn(u8g_t *u8g, u8g_dev_t *dev, uint8_t msg, void *arg)
+uint8_t u8g_dev_rrd_st7920_128x64_fn(u8g_t *u8g, u8g_dev_t *dev, uint8_t msg, void *arg)  // LGF changed int from uint8_t
 {
-  uint8_t i,y;
+  int i,y;
   switch(msg)
   {
     case U8G_DEV_MSG_INIT:
